Add `Validator` middleware. `Validator` ensures that all arguments for an event are valid. This currently only affects arguments that are `Shark::Object` instances; they will be validated against their respective schemas.
Only events that have all valid arguments will be proxied through to the next middleware, ensuring that only valid events are published.

+++ objects/vehicle.rb
@@ -15,7 +15,7 @@ module Shark
     attribute :onboard,         type: Integer,  nilable: true
     # [Float] The fullness of the vehicle expressed as a percentage in the
     # range [0-1]
-    attribute :saturation,      type: Float,    nilable: true
+    attribute :saturation,      type: Numeric,  nilable: true
     # [String] The last stop that this vehicle departed from
     # NOTE: Only the identifier will be stored. If more information is needed,
     # a lookup must be performed on the storage adapter.
